+++ cpp/locality/NeighborQuery.h
@@ -81,7 +81,18 @@ public:
     //! Constructor
     NeighborQuery(const box::Box& box, const vec3<float>* points, unsigned int n_points)
         : m_box(box), m_points(points), m_n_points(n_points)
-    {}
+    {
+        if (m_box.is2D())
+        {
+            for (unsigned int i(0); i < n_points; i++)
+            {
+                if (m_points[i].z != 0)
+                {
+                    throw std::invalid_argument("A point with z != 0 was provided in a 2D box.");
+                }
+            }
+        }
+    }
 
     //! Empty Destructor
     virtual ~NeighborQuery() {}