+++ ChangeLog.md
@@ -32,6 +32,7 @@ and this project adheres to
 * Steinhardt uses query arguments.
 * APIs for several order parameters have been standardized.
 * SolidLiquid order parameter has been completely rewritten, fixing several bugs and simplifying its C++ code.
+* NeighborQuery objects require z == 0 for all points if the box is 2D.
 
 ### Fixed
 * Steinhardt uses the ThreadStorage class and properly resets memory where needed.